--- homepage/app/_components/index.jsx
@@ -1,4 +1,0 @@
-// components export
-
-export { default as Header } from "./Header";
-export { default as Policy } from "./Policy";
